# backend/core-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse:
    properties:
      cluster:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateClusterRequest:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateDropletRequest:
    properties:
      cluster_id:
        description: Optional cluster ID to assign this droplet to
        type: string
      image:
        example: ubuntu-20-04-x64
        type: string
      name:
        example: web-01
        type: string
      region:
        example: nyc3
        type: string
      size:
        example: s-1vcpu-1gb
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateJobRequest:
    properties:
      parameters:
        additionalProperties:
          type: string
        type: object
      type:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteClusterResponse:
    properties:
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteDropletResponse:
    properties:
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterRequest:
    properties:
      cluster_id:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterResponse:
    properties:
      cluster:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
      insights:
        items:
          type: string
        type: array
      recommendations:
        items:
          type: string
        type: array
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DropletResponse:
    properties:
      droplet:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Droplet'
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse:
    properties:
      error:
        description: |-
          The error message
          example: Invalid request payload
        type: string
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloRequest:
    properties:
      name:
        example: Ava
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloResponse:
    properties:
      message:
        example: Hello, Ava from ClusterGenie!
        type: string
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.JobResponse:
    properties:
      job:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Job'
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListClustersResponse:
    properties:
      clusters:
        items:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
        type: array
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListDropletsResponse:
    properties:
      droplets:
        items:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Droplet'
        type: array
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListJobsResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Job'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.UpdateClusterRequest:
    properties:
      name:
        type: string
      region:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for ClusterGenie - DigitalOcean droplet management and monitoring
  title: ClusterGenie API
  version: "1.0"
paths:
  /clusters:
    get:
      consumes:
      - application/json
      description: Returns a list of clusters
      produces:
      - application/json
      responses:
        "200":
          description: List of clusters
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListClustersResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List clusters
      tags:
      - clusters
    post:
      consumes:
      - application/json
      description: Create a new cluster entity
      parameters:
      - description: Create cluster request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cluster created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Create a new cluster
      tags:
      - clusters
  /clusters/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the cluster and returns a confirmation message
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteClusterResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Delete cluster
      tags:
      - clusters
    get:
      consumes:
      - application/json
      description: Retrieve a cluster by its id
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster retrieved
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get cluster by id
      tags:
      - clusters
    put:
      consumes:
      - application/json
      description: Update an existing cluster's fields
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      - description: Update cluster request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated cluster
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Update cluster
      tags:
      - clusters
  /diagnosis/diagnose:
    post:
      consumes:
      - application/json
      description: Run a diagnosis on the supplied cluster and get insights/recommendations
      parameters:
      - description: Diagnose request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Diagnosis result
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Diagnose cluster
      tags:
      - diagnosis
  /droplets:
    get:
      consumes:
      - application/json
      description: Returns a list of droplets (optionally filtered by cluster)
      produces:
      - application/json
      responses:
        "200":
          description: List of droplets
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListDropletsResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List droplets
      tags:
      - droplets
    post:
      consumes:
      - application/json
      description: Provisions a new droplet. Provide region/size/image and optionally
        a cluster_id
      parameters:
      - description: Create droplet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateDropletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Droplet created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DropletResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error while provisioning
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Create a new droplet
      tags:
      - droplets
  /droplets/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a droplet and returns a confirmation message
      parameters:
      - description: Droplet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteDropletResponse'
        "404":
          description: Droplet not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Delete droplet
      tags:
      - droplets
    get:
      consumes:
      - application/json
      description: Returns a single droplet by id
      parameters:
      - description: Droplet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Droplet retrieved
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DropletResponse'
        "404":
          description: Droplet not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get droplet by id
      tags:
      - droplets
...
  /hello:
    post:
      consumes:
      - application/json
      description: Returns a greeting message for the supplied name
      parameters:
      - description: Hello request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Greeting created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Say hello
      tags:
      - hello
  /jobs:
    get:
      consumes:
      - application/json
      description: Returns paginated jobs list with sorting options
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort column
        in: query
        name: sort_by
        type: string
      - description: Sort direction (asc/desc)
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of jobs
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListJobsResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create a background job (provision/diagnose/scale/monitor)
      parameters:
      - description: Create job request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Job created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.JobResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Create job
      tags:
      - jobs
  /jobs/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve job status/result by id
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job retrieved
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.JobResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get job by id
      tags:
      - jobs
...
  /observability/ratelimit:
    get:
      consumes:
      - application/json
      description: Get available tokens and configuration for a rate limiter (supports
        scope filters)
      parameters:
      - description: Limiter name (e.g. diagnosis or jobs_create)
        in: query
        name: name
        required: true
        type: string
      - description: Optional scope type (global/user/cluster)
        in: query
        name: scope_type
        type: string
      - description: Optional scope id for user/cluster
        in: query
        name: scope_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Limiter status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing/invalid input
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "404":
          description: No such limiter
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Query rate limiter status
      tags:
      - observability
  /observability/ratelimit/config:
    delete:
      consumes:
      - application/json
      description: Delete a stored limiter config by key or name+scope
      parameters:
      - description: Delete config request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Deletion result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing input
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Delete persisted limiter config
      tags:
      - observability
    get:
      consumes:
      - application/json
      description: Retrieve persisted limiter config for a name and optional scope
      parameters:
      - description: Limiter name
        in: query
        name: name
        required: true
        type: string
      - description: Optional scope type
        in: query
        name: scope_type
        type: string
      - description: Optional scope id
        in: query
        name: scope_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Found config
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing name
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get persisted limiter config
      tags:
      - observability
    post:
      consumes:
      - application/json
      description: Store or update refill/capacity for a named limiter and scope
      parameters:
      - description: Limiter config request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Config saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Persist limiter configuration
      tags:
      - observability
  /observability/ratelimit/config/list:
    get:
      consumes:
      - application/json
      description: Returns list of stored limiter configs filtered by name/scope
      parameters:
      - description: Optional limiter name filter
        in: query
        name: name
        type: string
      - description: Optional scope type
        in: query
        name: scope_type
        type: string
      - description: Optional scope id
        in: query
        name: scope_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of configs
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List persisted limiter configs
      tags:
      - observability
...
swagger: "2.0"